Address review on the signed-extension install UI
Eight findings from #166. The three that were defects rather than tidying:

The signing payload claimed to be length-prefixed and was newline-terminated,
so a manifest whose field value contained a newline could reproduce another
manifest's byte sequence and reuse its signature. Now genuinely
length-prefixed: `<len>:<name><len>:<value>`.

The permission review emitted one row per requested-capability *entry*, and
`index_of` resolves a capability to the first matching row. A manifest that
listed a capability twice therefore got a review the user could not complete
(the second row was undecidable) that approved anyway, and a denial recorded
on the duplicate row was discarded. Rows and grants are now deduplicated, with
a per-row-index denial test -- the first draft of that test passed with the
dedup disabled, because deciding by capability reached the first row either way.

`artifact_resigned_by_an_untrusted_key_is_refused` kept
`signer = "legion-first-party"`, which is anchored, so it exercised the
signature check with the wrong key and never the unknown-signer guard it was
named for. Renamed to what it tests; the untrusted-signer guard now has its own.

Also: the extensions panel projected no accessibility node, so the one surface
that asks a user to grant capabilities to third-party code was silent to a
screen reader -- it now announces its count and pending reviews, which is what
`audit_rows()` claimed to feed and did not (deleted; its two test uses now
assert typed fields). `ExtensionCatalogError::Registry` carries the registry's
typed refusal instead of its rendered text. `is_installable` collapsed a typed
refusal into a bool and had no caller outside tests. The unused `#[from]` on
`PermissionReview` is now `#[source]`. `ExtensionPermissionProjection.ordinal`
is `u32`, not a host-width `usize`, on a type that serializes.

+++ crates/legion-protocol/src/extensions.rs
@@ -100,7 +100,11 @@ impl ExtensionPermissionState {
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct ExtensionPermissionProjection {
     /// 1-based position in the review list.
-    pub ordinal: usize,
+    ///
+    /// `u32`, not `usize`: this crosses the projection boundary as serialized
+    /// data, and a width that depends on the host it was encoded on is not a
+    /// wire type. No review has 4 billion rows.
+    pub ordinal: u32,
     /// Capability this row grants or withholds.
     pub capability: CapabilityId,
     /// Short human-readable title.
